+++ example/lib/main.dart
@@ -62,9 +62,11 @@ class _HomeViewState extends State<HomeView> {
                 final String? response =
                     await FlutterLivenessDetectionRandomizedPlugin.instance
                         .livenessDetection(
+                  showCurrentStep: true,
                   context: context,
                   shuffleListWithSmileLast: false,
                   isEnableSnackBar: true,
+                  isDarkMode: false,
                   config: LivenessDetectionConfig(
                     startWithInfoScreen: true,
                   ),