+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3 🚀
+
+- 🛠️ Adjust to compatible camera dependency to prevent face not found
+- 🔐 Ajdust threshold for smile and look down challenge
+- 🎨 Add showCurrentStep parameter (default : false)
+- 🎨 Add Light and Dark mode
+
 ## 1.0.2 🚀
 
 ### Update README.md